backend/requirements.txt
ddtrace
google-adk
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic[email]
python-jose[cryptography]
